Move ROS related functions into separate file
