+++ rt_gene/src/rt_gene/__init__.py
@@ -1,6 +1,7 @@
 from . import estimate_gaze_base
 from . import extract_landmarks_method_base
 from . import gaze_tools
+from . import ros_tools
 from . import tracker_generic
 # from . import tracker_face_encoding
